+++ test/specs/basic/array-keyword.js
@@ -22,7 +22,6 @@ describe('basic mark with an array of keywords', () => {
     expect($ctx.find('mark').length).toBe(8);
   });
   it('should call noMatch for not found array items', () => {
-    console.log( notFound );
     expect(notFound).toEqual(['test', 'hey']);
   });
 });